hotfix

+++ pointmar/diffusion/respace.py
@@ -122,8 +122,13 @@ class _WrappedModel:
         self.original_num_steps = original_num_steps
 
     def __call__(self, x, ts, **kwargs):
-        map_tensor = th.tensor(self.timestep_map, device=ts.device, dtype=ts.dtype)
-        new_ts = map_tensor[ts]
+        # Ensure integer indexing: timestep tensors may be float (fractional timesteps
+        # are used in some embeddings). Cast to long and build the mapping tensor
+        # as a long tensor so indexing works reliably.
+        # map_tensor = th.tensor(self.timestep_map, device=ts.device, dtype=ts.dtype)
+        # new_ts = map_tensor[ts]
+        map_tensor = th.tensor(self.timestep_map, device=ts.device, dtype=th.long)
+        new_ts = map_tensor[ts.long()]
         # if self.rescale_timesteps:
         #     new_ts = new_ts.float() * (1000.0 / self.original_num_steps)
         return self.model(x, new_ts, **kwargs)
